+++ packages/camera/camera_avfoundation/example/ios/RunnerTests/Mocks/MockCamera.swift
@@ -27,6 +27,7 @@ final class MockCamera: NSObject, Camera {
   var lockCaptureOrientationStub: ((PlatformDeviceOrientation) -> Void)?
   var unlockCaptureOrientationStub: (() -> Void)?
   var setImageFileFormatStub: ((PlatformImageFileFormat) -> Void)?
+  var setImageQualityStub: ((Int64) -> Void)?
   var setExposureModeStub: ((PlatformExposureMode) -> Void)?
   var setExposureOffsetStub: ((Double) -> Void)?
   var setExposurePointStub: ((PlatformPoint?, @escaping (Result<Void, any Error>) -> Void) -> Void)?
@@ -142,6 +143,10 @@ final class MockCamera: NSObject, Camera {
 
   func setImageFileFormat(_ fileFormat: PlatformImageFileFormat) {
     setImageFileFormatStub?(fileFormat)
+  }
+
+  func setImageQuality(_ quality: Int64) {
+    setImageQualityStub?(quality)
   }
 
   func setExposureMode(_ mode: PlatformExposureMode) {