+++ packages/camera/camera_android_camerax/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.7.2
+
+* Adds `setImageQuality` for controlling JPEG compression quality.
+
 ## 0.7.1
 
 * Removes outdated restrictions against concurrent camera use cases.